Story 7  (Final for user story 7) Added New pet tab Set tab to second column Program MUST be restarted for values to be taken into memory. While they are added to the CSV I was unable to get the program to dump and refersh it's array and tables
The following would have to happen for this to occur

Add a line to the pet array of values dump and refresh

Unless the tables were to refresh automatically the following would also have to occur:
Run an add line from array to table this would have to occur on the main table and the table of the shop which was chosen.
For the shop table take the shop which was selected then find and add to the table of the shop's same name use a loop to research the tab for first the name of the shop then add to its' table using an add row method.

For the shop dump the data and refersh it. This table has calculations which need to be recalculated upon the addition of a new price value. It would be easier to clear. Don't delete the tab just the values in the table and then re add them